+++ pkg/services/container/announcement/load/storage/storage.go
@@ -40,21 +40,22 @@ type storageKey struct {
 type Storage struct {
 	mtx sync.RWMutex
 
-	mItems map[storageKey]*usedSpaceEstimations
+	estLifeCycle uint64
+	mItems       map[storageKey]*usedSpaceEstimations
 }
-
-// Prm groups the required parameters of the Storage's constructor.
-//
-// The component is not parameterizable at the moment.
-type Prm struct{}
 
 // New creates a new instance of the Storage.
 //
 // The created Storage does not require additional
 // initialization and is completely ready for work.
-func New(_ Prm) *Storage {
+//
+// estimationsLifeCycle is a longevity (in epochs) of estimations
+// that are kept in the [Storage] instance. Note, current epoch
+// is controlled with [Storage.EpochEvent].
+func New(estimationsLifeCycle uint64) *Storage {
 	return &Storage{
-		mItems: make(map[storageKey]*usedSpaceEstimations),
+		mItems:       make(map[storageKey]*usedSpaceEstimations),
+		estLifeCycle: estimationsLifeCycle,
 	}
 }
 
@@ -116,6 +117,20 @@ func (s *Storage) Iterate(f loadcontroller.UsedSpaceFilter, h loadcontroller.Use
 	s.mtx.RUnlock()
 
 	return
+}
+
+// EpochEvent notifies [Storage] about epoch counter updating.
+// Used to remove unused estimations. See [Prm.EstimationsLifeCycle].
+// Blocking operation.
+func (s *Storage) EpochEvent(e uint64) {
+	s.mtx.Lock()
+	defer s.mtx.Unlock()
+
+	for k := range s.mItems {
+		if k.epoch+s.estLifeCycle < e {
+			delete(s.mItems, k)
+		}
+	}
 }
 
 func finalEstimation(vals []uint64) uint64 {